requirements.txt
fuzzywuzzy==0.18.0
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
PyYAML==6.0.1
psycopg2-binary==2.9.9
numpy==1.26.4
tensorflow==2.16.1
shap==0.45.0
flask==3.0.3
python-Levenshtein==0.25.1
streamlit==1.32.1
matplotlib==3.8.3
seaborn==0.13.2
lime==0.2.0.1